Security question collection
Created viewmodel for data. Created custom validataion attribute for checking that they select different options, the validation is done after they submitt. Added default selection that is disabled, so it is not concidered a valid option. Added security question action on the user controller to needs to store the data in the database.  Changed the database schema for users to include security questions.