Completed and working
Added 5 reset password methods. ResetPasswordConfirmedAsync forwards the user to a page where they can enter their new password. ResetPasswordAsync verifies the new password and makes the change, redirecting the user to a confirmed page. RequestResetPasswordConfirmation is the first set of methods that handle getting the users email.

The user controller does not manage login so the forget password method was removed.