Change Password Finished
Now when the user tries to change their password it will change in the database. Added ChangePasswardViewModel which stores the form data when the http post is submitted. Changed the ChangePassword.cshtml to use the new view model. Changed the user controller to have a change passward async function that touches the user database.