Database setup
The zip_code_database.csv is the raw data of the zip codes for the united states. The ZipCodeImportCreator.py is a script that reads the raw zip code data and then chooses which data to include and caculates the Voronoi diagram for the zip codes. The ZipCodeImport.sql is an SQL query to setup the ZipCode table in the database.