Imported changes made in the Admin/Manager branch.
Admin/Manager branch was too far from the master to merge, instead code was copied.
I added some major changes. Accesses the database to poulate the table. Delete and Confirm-Delete buttons are linked together and call commands on the server but the backend ignores the request. Edit calls commands on the server but the backend ignores the request.
Added styles to the table. Used danger buttons for the delete and confirm delete.

Co-Authored-By: honestly7 <honestly7@users.noreply.github.com>