Fixing some of the Admin functions
Admin can now add office positions and candidates. The dashboard of the general user reduced to only contain register to vote and enter demographics. Soon the admin dashboard link will be removed.
The database now has tables for offices and candidates. Added datamodels for each.
When adding a new candidate the office is provided as a select option.

Fixed a bug with voter registration and enter demographics when the user didn't have any yet.