Candidate setup
Add candidate allows the user to filter the users till they find the one they are looking for and filter the ballots in the same manner. Top 5 results are shown. Split the two forms in half using HTML 5 syntax.  Added districts, still in basic form without a proper layout. Zip codes not stay shown if they are selected.