Fix Create User and Email Confirmation
The error for user creation did not add the user to the "GenericUser" role, this was fixed by adding the user to that role after they confirm their email. The error was resending an email confirmation required the email to be confirmed, fixed in verify controller.
Reset password will confirm the user email and add them to the "GenericUser" role.