View Ballot
View Ballot view component provides the action button when viewing a list of ballots.
View ballot function in the admin controller sets up the model to display a ballot.
searchballot provides a list of ballots based on the fileters set in the basic search view model.
The view ballot view model is all the properties that we can show in the browser.
search ballot is for displaying a list of ballots that can be filtered.